chore: move config into ts wrapper
