+++ src/config.ts
@@ -1,0 +1,3 @@
+import {readFileSync} from "fs";
+
+export default JSON.parse(readFileSync("config.json").toString())
